backend/requirements.txt
uvicorn==0.34.0
backend==0.2.4.1
cryptography==44.0.2
fastapi==0.115.12
openai==1.76.0
passlib==1.7.4
pydantic==2.11.3
pytest==8.3.5
python-dotenv==1.1.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.9
python-jose==3.4.0
